+++ run-test.sh
@@ -6,9 +6,11 @@ cargo build --release --target-dir target
 cd ../..
 
 # build test hApp
-cd ts/test/fixture/zomes/entry
+cd ts/test/fixture/zomes/integrity
 cargo build --release --target wasm32-unknown-unknown
-cd ../../ # into fixtures
+cd ../coordinator
+cargo build --release --target wasm32-unknown-unknown
+cd ../.. # into fixtures
 hc dna pack . -o entry.dna
 hc app pack . -o entry.happ
 cd ../../..