+++ CHANGELOG.md
@@ -9,6 +9,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 ### Fixed
 
+## [0.6.1]
+
+### Changed
+- Fixture: split zome into integrity and coordinator zomes [PR \#143](https://github.com/holochain/tryorama/pull/143)
+
+## [0.6.0]
+
+### Changed
+- Holochain: Upgrade Holochain & @holochain/client & replace Element/Header with Action/Record [PR \#142](https://github.com/holochain/tryorama/pull/142)
+
 ## [0.5.9]
 
 ### Added